Module Vcs_git

Implementation of a git provider for the Vcs library, based on Eio and Git_cli.

This implementation is based on the git command line tool. We run it as an external program within an Eio environment, producing the right command line invocation and parsing the output to produce a typed version of the expected results with Git_cli. Note that git must be found in the PATH of the running environment.

type 'a t = [> Git_cli.Trait.t ] as 'a Vcs.t

type t' = Git_cli.Trait.t t

This is a convenient type alias that may be used to designate a provider with the exact list of traits supported by this implementation.

val create : env:< fs : _ Eio.Path.t ; process_mgr : _ Eio.Process.mgr.. > -> _ t

create ~env creates a vcs value that can be used by the Vcs library.

module Impl : sig ... end

The implementation of the provider is exported for convenience. Casual users should prefer using Vcs directly.